requirements.txt
Flask==2.3.3
pymongo==4.13.2
python-dotenv==1.0.0
marshmallow==3.20.1
flask-cors==4.0.0
APScheduler==3.10.4
pytest==7.4.3
pytest-flask==1.3.0